Expect API at /api, proxy to 3001 in development
In production scenarios, it is not ideal for the browser to communicate to the API via a different server or port.  Instead, it should make requests to /api over the same port as the web app.  This requires proxying /api to port 3001.

We support this via the basic approach outlined here:
https://github.com/facebookincubator/create-react-app/blob/master/packages/react-scripts/template/README.md#proxying-api-requests-in-development

This allows the app's JS to be configured for production deployment (sending API calls to /api).

In development (when started via `yarn run start-dev`), a simple proxy is utilized to proxy any unresolved paths (e.g., /api) to port 3001.  This should be transparent to developers.

In the docker containers, by default, the fronted.js will also provide a simple express-based proxy so things work as expected.  This can be disabled via an environment variable, however, when a production environment uses an nginx or Apache httpd to proxy /api to the API server (so the express-based proxy is not needed).