Disable Presets
Presets are a little buggy, and probably don't represent how we'll support this capability in the future anyway.  So... we're removing them before release so we don't confuse users -- and so we don't need to worry about preserving presets if we design a better way forward.

This just comments them out in the JSX and updates the docs (mostly screenshots).  In the future we'll decided whether or not we completely rip them out.