Add polyfills for IE 11
The create-react-app scripts compile for "modern browsers", which does not include IE 11.  To address this, we need to polyfill elements that we use but IE 11 does not support.  This adds basic polyfills for commonly used things -- but we may need to add more in the future if/when we use other things IE 11 doesn't support.

There are more robust solutions using babel presets, but unfortunately create-react-app doesn't allow you to control them unless you "eject".