Reduce likelihood of time mismatch in artifacts test
The artifacts test asserts equality on status messages that contain time stamps down to the second.  Formerly, a static time was captured when the artifacts.test.js file was loaded, and tests used it as the _expected_ time stamp.  If the clock moved to the next second between when the file was loaded and when the test was executed, the tests would fail since the timestamp changed.

This commit causes each test to get the _expected_ timestamp right before it executes the code -- greatly reducing the likelihood of a timestamp mismatch.  That said, a mismatch is still possible -- but coming up with a more robust solution would be a good bit uglier.