Don't try to detect FHIR Auth Type
Since we know NLM uses Basic Auth, and we do not have immediate plans to support other servers, there is no need to detect the auth type.

While it's generally harmless, this is being removed because the NLM FHIR server is now returning HTTP 500 when querying its metadata -- and this is causing logins to fail.  Since it's totally unnecessary, it's better and easier just to remove it.