Update CDS Connect Commons for FHIR R4
Update the R4 commons files to add back in functions that may not be used by the AT, but are used by other libraries that also use the commons file.  We want to ensure that the CDS Connect Commons used by the AT is the same CDS Connect Commons used by others.