CodeSelectModal: Replace ICD-10 w/ ICD-10-CM, Rename ICD-9 to ICD-9-CM
The code select modal shouldn't allow users to choose base ICD-10, as VSAC doesn't support it.  So switch it to ICD-10-CM.

ICD-9 already pointed to the ICD9CM URL, but to make it more clear, rename it to ICD-9-CM.

This does not affect existing artifacts.  If they used ICD-10, they will retain the base ICD-10 unless/until an author removes the ICD-10 code and adds an ICD-10-CM code.