Fix naming clashes in library includes when external CQL has CDS Connect Commons libraries
When external CQL is added, all external CQL libraries are automatically added to the artifact as "includes".  This causes a problem if CDS Connect Commons libraries are uploaded, because the artifact already has includes defined for them.  We then end up w/ two sets of includes per commons library -- one with a localIdentifier (C3F) and one without.  This then messes up cql-merge which keeps a map of library names to local identifiers.

To fix this, rename the CDS Connect Commons that we internally use by prefixing AT_Internal to the name.  That way there are no clashes and cql-merge works properly.