Merged in cmm-bug-fixing-round-1 (pull request #515)
Bug Fixapalooza

* Fix resource generation scripts to properly detect allowsCustomCodes

* Fix query build for DSTU2 MedicationOrder

Even though it is a MedicationOrder in FHIR DSTU2, the return type is still list_of_medication_requests -- so we need to account for that when querying for resource data.

* Quantity Comparison operator tweaks

- Use words (e.g., "greater than") instead of symbols (e.g. ">")
- Label the comparison as "Comparison"

* Use year(s)/month(s)/... instead of years/months/...

This also required a change in the expression generator so that it doesn't strip out the ( ) characters.

* Tweak postLabels for readability

- elements -> element(s)
- remove "old" postlabel for age operators

* Remove unnecessary space in dateTimeTypeWithinLast CQL template

* Rename and fix occurred within / more than operators

- Operator names were not consistent w/ similar operators
- Duration should be an integer (not decimal)
- Date fields should not get options for hours/minutes, so they need their own operators
- Date fields also need to use Today() in CQL instead of Now()

* Handle instant/dateTime in dateTimeOccursBetween when arguments are dates only

* Remove support for System.Date in dateTimeOccurred and dateTimeOccursBetween

CQL doesn't like it when you use this with a date property but specified arguments that are DateTimes (with time components). The long-term solution is to support a date-only editor, but for now we'll disable these operators.

* Fix CQL download for MedicationOrder w/ custom modifier


Approved-by: Peter Weeks
Approved-by: Noranda Brown