Merged in fhir401 (pull request #527)
Support for FHIR 4.0.1

* Update Included CQL Libraries

Update FHIRHelpers base on versions in the CQL 1.5.6 release:
https://github.com/cqframework/clinical_quality_language/releases/tag/v1.5.6

Bump version number in R4 CDS Connect Commons (which should have been bumped a while …
* Rename CQL folders to match fhir version number

Since R4 could be 4.0.0 or 4.0.1, we need to distinguish CQL folders by the number (e.g., 4.0.0) rather than the label (e.g., R4).

* Add support for FHIR 4.0.1

This adds support for FHIR 4.0.1 in addition to FHIR 4.0.0. Users can choose to download either one, unless an external CQL library locks it down to a specific version.

This also fixes a bug in artifact upload that was caused b…
* Add CQL Handler tests for 4.0.x, 4.0.1, and 4.0.0.

* Migration to change 4.0.0 to 4.0.x (when appropriate)

- If existing artifact is version 4.0.0, but has no external CQL, it can safely be 4.0.x
- Otherwise, if existing artifact is 4.0.0 w/ external artifacts, keep it 4.0.0

* Fix modifier handlers and CQL execution for 4.0.1

* Note availability of 4.0.1 in Docs

* Add FHIR 4.0.1 to What's New

* Add additional functions to FHIR 4.0.1 Commons library

- Update library and bump library version
- Update includedLibraries version. Note that these versions
  currently aren't used in the generated CQL because they
  are stripped after cql-merge merges.

* Patch fix for a minor version


Approved-by: Julia Afeltra