Merged in simplify-docker-build (pull request #586)
Simplify docker build

* Remove testing from Docker build

Testing inside the Docker container caused issues on some systems, as it used a lot of memory and ran very long.

This commit removes the testing components from the container so that the build is focused only on building what is needed to run in production. It is assumed that tests have been run prior to doing the official build.

* Remove husky

The husky tool has caused problems in different local configurations and is no longer as necessary since we run tests in a Bitbucket pipeline.

* Remove commented out line in Dockerfile


Approved-by: Julia Afeltra