Merged in fix-docker-cp (pull request #593)
Docker: When copying files to folders, put trailing / on folder

The version of Docker in the build environment requires you to put a slash at the end of the folder name when copying multiple files to a folder. Otherwise you get this error: "When using COPY with more than one source file, the destination must be a directory and end with a /".