Merged in migrate-cql-merge (pull request #633)
Migrate cql-merge code to api

The cql-merge code that was developed by Dylan Mahalingam previously existed only as an unpublished module, stored in localDependencies as a package tgz. In order to improve maintainability, the cql-merge code was converted from typescript to javascript and migrated to the existing api module here.

Approved-by: Julia Afeltra