Merged in api-dep-updates-june24 (pull request #636)
API Dependency Updates for June

* Update patch versions of dependencies

* Update minor versions of dependencies except antlr4

Treat anltr4 like a major version update because it does not follow semantic versioning.

* Re-generate ANTLR4 classes with ANTLR4 4.8.0 and add dependency note

Since ANTLR4 4.8.0 is the last version that supports Common JS, that is the latest version we can use.

* Update major versions of dependencies

* npm audit fix

* Remove uuid from api dependencies

The uuid package is only used for the old migration 17, which is no longer active code.


Approved-by: Julia Afeltra